fix(ci): split hosted runtime bundles
