+++ packages/internal/package.json
@@ -16,7 +16,9 @@
     "./nitro": "./src/nitro.ts",
     "./object": "./src/object.ts",
     "./runtime/app": "./src/runtime/app.ts",
+    "./runtime/cloudflare-hosted": "./src/runtime/cloudflare-hosted.ts",
     "./runtime/hosted": "./src/runtime/hosted.ts",
+    "./runtime/vercel-hosted": "./src/runtime/vercel-hosted.ts",
     "./runtime/cloudflare-env": "./src/runtime/cloudflare-env.ts",
     "./runtime/random": "./src/runtime/random.ts",
     "./build/cloudflare": "./src/build/cloudflare.ts",
